+++ NEWS.md
@@ -1,12 +1,12 @@
 # segtest 2.0.0
 
-* Added a `NEWS.md` file to track changes to the package.
-* `seg_lr()`: Tests for segregation distortion for arbitrary (even) ploidy F1 populations.
+* `seg_lrt()`: Tests for segregation distortion for arbitrary (even) ploidy F1 populations.
+* `seg_multi()`: Runs `seg_lrt()` at many loci, with parallelization support from the future package.
 * `gamfreq()`: Gamete frequencies under various models of meiosis.
 * `gf_freq()`: Genotype frequencies of an F1 population under various models of meiosis.
 * `drbounds()`: Two different upper bounds on the double reduction rates.
 * `beta_bounds()`: Two different upper bounds on double reduction adjustment for simplex loci.
 * `n_pp_mix()`: Number of mixing components in the pairing configuration model for partial preferential pairing.
-
+* Added a `NEWS.md` file to track changes to the package.